OS-21 add hibernate config and migration provider

+++ owl-impl/src/main/java/com/geo/owl/building/model/BuildingObject.java
@@ -40,5 +40,5 @@ public class BuildingObject {
     private Set<WorkArea> workAreas = new HashSet<>();
 
     @Version
-    private int lock;
+    private long lock;
 }
